refactor(knowledge): split facts.py into ingestion/entries/query schema modules (#5486)
Move 13 Pydantic response schemas from the monolithic facts.py into three
focused sub-modules: ingestion.py (add/upload/clear responses), entries.py
(stored entry retrieval shapes), and query.py (search/query responses).
facts.py is now a backward-compat re-export shim so all existing callers
continue to work without modification.
